+++ spec/client_spec.rb
@@ -1,57 +1,73 @@
-# require 'spec_helper'
+require 'spec_helper'
 
-# describe 'percona::client' do
-#   describe 'when `package_action` is `install`' do
-#     describe 'Ubuntu' do
-#       let(:chef_run) do
-#         ChefSpec::SoloRunner.new.converge(described_recipe)
-#       end
+describe 'percona::client' do
+  before do
+    stub_command('dnf module list mysql | grep -q "^mysql.*\\[x\\]"')
+  end
 
-#       specify do
-#         expect(chef_run).to install_package 'libperconaserverclient18.1-dev'
-#         expect(chef_run).to install_package 'percona-server-client-5.6'
-#       end
-#     end
+  describe 'when `package_action` is `install`' do
+    context 'Ubuntu' do
+      platform 'ubuntu'
 
-#     describe 'CentOS' do
-#       let(:chef_run) do
-#         env_options = { platform: 'centos', version: '6' }
-#         ChefSpec::SoloRunner.new(env_options).converge(described_recipe)
-#       end
+      it do
+        expect(chef_run).to install_package 'percona-server-client'
+      end
 
-#       specify do
-#         expect(chef_run).to install_package 'Percona-Server-devel-56'
-#         expect(chef_run).to install_package 'Percona-Server-client-56'
-#       end
-#     end
-#   end
+      context 'version 5.7' do
+        override_attributes['percona']['version'] = '5.7'
+        it do
+          expect(chef_run).to install_package 'percona-server-client-5.7'
+        end
+      end
 
-#   describe 'when `package_action` is `upgrade`' do
-#     describe 'Ubuntu' do
-#       let(:chef_run) do
-#         ChefSpec::SoloRunner.new do |node|
-#           node.default['percona']['client']['package_action'] = 'upgrade'
-#         end.converge(described_recipe)
-#       end
+      context 'version 5.6' do
+        override_attributes['percona']['version'] = '5.6'
+        it do
+          expect(chef_run).to install_package 'percona-server-client-5.6'
+        end
+      end
+    end
 
-#       specify do
-#         expect(chef_run).to upgrade_package 'libperconaserverclient18.1-dev'
-#         expect(chef_run).to upgrade_package 'percona-server-client-5.6'
-#       end
-#     end
+    context 'CentOS' do
+      platform 'centos'
 
-#     describe 'CentOS' do
-#       let(:chef_run) do
-#         env_options = { platform: 'centos', version: '6' }
-#         ChefSpec::SoloRunner.new(env_options) do |node|
-#           node.default['percona']['client']['package_action'] = 'upgrade'
-#         end.converge(described_recipe)
-#       end
+      it do
+        expect(chef_run).to install_package 'percona-server-client'
+      end
 
-#       specify do
-#         expect(chef_run).to upgrade_package 'Percona-Server-devel-56'
-#         expect(chef_run).to upgrade_package 'Percona-Server-client-56'
-#       end
-#     end
-#   end
-# end
+      context 'version 5.7' do
+        override_attributes['percona']['version'] = '5.7'
+        it do
+          expect(chef_run).to install_package 'Percona-Server-client-57'
+        end
+      end
+
+      context 'version 5.6' do
+        override_attributes['percona']['version'] = '5.6'
+        it do
+          expect(chef_run).to install_package 'Percona-Server-client-56'
+        end
+      end
+    end
+  end
+
+  describe 'when `package_action` is `upgrade`' do
+    context 'Ubuntu' do
+      platform 'ubuntu'
+      override_attributes['percona']['client']['package_action'] = 'upgrade'
+
+      it do
+        expect(chef_run).to upgrade_package 'percona-server-client'
+      end
+    end
+
+    context 'CentOS' do
+      platform 'centos'
+      override_attributes['percona']['client']['package_action'] = 'upgrade'
+
+      it do
+        expect(chef_run).to upgrade_package 'percona-server-client'
+      end
+    end
+  end
+end