+++ spec/default_spec.rb
@@ -3,10 +3,6 @@ require 'spec_helper'
 describe 'percona::default' do
   platform 'ubuntu'
 
-  before do
-    stub_command('apt-key list | grep 8507EFA5')
-  end
-
   it do
     expect(chef_run).to include_recipe('percona::client')
   end