+++ spec/package_repo_spec.rb
@@ -2,18 +2,44 @@ require 'spec_helper'
 
 describe 'percona::package_repo' do
   before do
-    stub_command('apt-key list | grep 8507EFA5')
     stub_command('dnf module list mysql | grep -q "^mysql.*\\[x\\]"')
   end
+
   context 'ubuntu' do
     platform 'ubuntu'
 
     it do
-      expect(chef_run).to add_apt_repository('percona')
+      expect(chef_run).to add_apt_repository('percona').with(
+        uri: 'http://repo.percona.com/apt',
+        components: %w(main),
+        keyserver: 'hkp://keys.gnupg.net:80',
+        key: %w(8507EFA5)
+      )
     end
 
     it do
-      expect(chef_run).to add_apt_preference('00percona')
+      expect(chef_run).to add_apt_preference('00percona').with(
+        glob: '*',
+        pin: 'release o=Percona Development Team',
+        pin_priority: '1001'
+      )
+    end
+
+    it do
+      expect(chef_run).to add_apt_repository('percona-ps-80').with(
+        uri: 'http://repo.percona.com/ps-80/apt',
+        components: %w(main),
+        keyserver: 'hkp://keys.gnupg.net:80',
+        key: %w(8507EFA5)
+      )
+    end
+
+    context 'version < 8.0' do
+      override_attributes['percona']['version'] = '5.7'
+
+      it do
+        expect(chef_run).to_not add_apt_repository('percona-ps-80')
+      end
     end
   end
 
@@ -22,19 +48,49 @@ describe 'percona::package_repo' do
 
     it do
       expect(chef_run).to create_yum_repository('percona').with(
+        description: 'Percona Packages',
+        baseurl: 'http://repo.percona.com/yum/release/$releasever/RPMS/$basearch',
         gpgkey: [
           'https://repo.percona.com/yum/PERCONA-PACKAGING-KEY',
           'https://repo.percona.com/yum/RPM-GPG-KEY-Percona',
-        ]
+        ],
+        gpgcheck: true,
+        sslverify: true
       )
     end
+
     it do
       expect(chef_run).to create_yum_repository('percona-noarch').with(
+        description: 'Percona Packages - noarch',
+        baseurl: 'http://repo.percona.com/yum/release/$releasever/RPMS/noarch',
         gpgkey: [
           'https://repo.percona.com/yum/PERCONA-PACKAGING-KEY',
           'https://repo.percona.com/yum/RPM-GPG-KEY-Percona',
-        ]
+        ],
+        gpgcheck: true,
+        sslverify: true
       )
+    end
+
+    it do
+      expect(chef_run).to create_yum_repository('percona-ps-80').with(
+        description: 'Percona Packages - ps-80',
+        baseurl: 'http://repo.percona.com/ps-80/yum/release/$releasever/RPMS/$basearch',
+        gpgkey: [
+          'https://repo.percona.com/yum/PERCONA-PACKAGING-KEY',
+          'https://repo.percona.com/yum/RPM-GPG-KEY-Percona',
+        ],
+        gpgcheck: true,
+        sslverify: true
+      )
+    end
+
+    context 'version < 8.0' do
+      override_attributes['percona']['version'] = '5.7'
+
+      it do
+        expect(chef_run).to_not create_yum_repository('percona-ps-80')
+      end
     end
   end
 end