+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of the percona cookbook.
 - Add support for CentOS 8
 - Add cluster suite and tests to test cluster recipe
 - Add support for Percona 8.0 and default to that version
+- Re-add ChefSpec tests
 
 ### Fixed